qa: stuff a bunch of stuff about argument shapes in test providers into the Psalm baseline
New version of Psalm is flagging most of the providers and/or inline input filter configuration as being invalid, when it is not.

Additionally, it does not like lazy-initialization via getters.
I have made the event manager and input filter nullable to solve one set of errors, but now Psalm is complaining that returning the property from a getter that does lazy initialization is invalid (it's not).

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>